+++ CHANGELOG.md
@@ -14,6 +14,10 @@
 
 * Support for custom font fallbacks
   (https://codeberg.org/dnkl/yambar/issues/153).
+* overline: new decoration
+  (https://codeberg.org/dnkl/yambar/issues/153).
+
+
 ### Changed
 ### Deprecated
 ### Removed